extract dot tests to a different file
