+++ yabt/graph_test.py
@@ -24,18 +24,14 @@ yabt target graph tests
 
 from concurrent.futures import ThreadPoolExecutor
 from functools import reduce
-import io
 import random
-import re
 
 import networkx
 import pytest
 
 from .buildcontext import BuildContext
-from .dot import write_dot, TARGETS_COLORS
 from .graph import (
     get_descendants, populate_targets_graph, topological_sort)
-from .extend import Plugin
 
 
 slow = pytest.mark.skipif(not pytest.config.getoption('--with-slow'),
@@ -312,80 +308,3 @@ def test_dep_name_typo(basic_conf):
     assert 'typo:zapi - dependency of typo:foo' in ex_msg
     # # expecting 5 unresolved targets (so error message will have 6 lines)
     assert 6 == len(ex_msg.split('\n'))
-
-
-@pytest.mark.usefixtures('in_dag_project')
-def test_graph_dot_generation(basic_conf):
-    build_context = BuildContext(basic_conf)
-    populate_targets_graph(build_context, basic_conf)
-    expected_targets = {':flask', ':gunicorn', 'common:logging', 'common:base',
-                        'fe:fe', 'yapi/server:users', 'yapi/server:yapi',
-                        'yapi/server:yapi-gunicorn'}
-    expected_dot_nodes = {'  "{}" [color="black",];'.format(target)
-                          for target in expected_targets}
-    expected_dot_edges = set([
-        '  "common:base" -> "common:logging";',
-        '  "fe:fe" -> "yapi/server:users";', '  "fe:fe" -> "common:base";',
-        '  "fe:fe" -> ":flask";', '  "yapi/server:yapi" -> "common:base";',
-        '  "yapi/server:yapi" -> ":flask";',
-        '  "yapi/server:yapi-gunicorn" -> "yapi/server:yapi";',
-        '  "yapi/server:yapi-gunicorn" -> "common:base";',
-        '  "yapi/server:yapi-gunicorn" -> ":gunicorn";'])
-    with io.StringIO() as dot_io:
-        write_dot(build_context, basic_conf, dot_io)
-        dot_lines = dot_io.getvalue().strip('\n').split('\n')
-        assert 'strict digraph  {' == dot_lines[0]
-        assert '}' == dot_lines[-1]
-        assert expected_dot_nodes == set(dot_lines[1:9])
-        assert expected_dot_edges == set(dot_lines[9:18])
-
-
-@pytest.mark.usefixtures('in_simpleflat_project')
-def test_graph_dot_generation_colors(basic_conf):
-    build_context = BuildContext(basic_conf)
-    populate_targets_graph(build_context, basic_conf)
-    expected_targets = {':flask-0.10.1': TARGETS_COLORS['PythonPackage'],
-                        ':flask-hello-app': TARGETS_COLORS['Python']}
-    expected_dot_nodes = ['  "{}" [color="{}",];'.format(target, color)
-                          for target, color in expected_targets.items()]
-    with io.StringIO() as dot_io:
-        write_dot(build_context, basic_conf, dot_io)
-        dot_lines = dot_io.getvalue().strip('\n').split('\n')
-        for dot_node in expected_dot_nodes:
-            assert dot_node in dot_lines
-
-
-@pytest.mark.usefixtures('in_cpp_project')
-def test_no_buildenv_deps_in_dot(basic_conf):
-    build_context = BuildContext(basic_conf)
-    basic_conf.targets = ['hello:hello-app']
-    populate_targets_graph(build_context, basic_conf)
-    buildenv_targets = {':builder', ':ubuntu-gpg', ':clang', ':ubuntu',
-                        ':gnupg'}
-    expected_targets = {'hello:hello-app', 'hello:hello'}
-    with io.StringIO() as dot_io:
-        write_dot(build_context, basic_conf, dot_io)
-        all_targets = set(dot_io.getvalue().split('"'))
-        assert not buildenv_targets.intersection(all_targets)
-        assert expected_targets.intersection(all_targets) == expected_targets
-
-
-@slow
-@pytest.mark.usefixtures('in_caching_project')
-def test_cached_targets_different_color(basic_conf):
-    basic_conf.targets = [':builder']
-    build_context = BuildContext(basic_conf)
-    populate_targets_graph(build_context, basic_conf)
-    cached_targets = {':build-tools', ':tools', ':unzip', ':ubuntu'}
-    other_targets = {':builder', ':builder-base'}
-    expected_dot_nodes = set([
-        '  "{}" \[color=".*",fillcolor="grey",style=filled\];'.format(target)
-        for target in cached_targets] + [
-        '  "{}" \[color=".*",\];'.format(target) for target in other_targets])
-    with io.StringIO() as dot_io:
-        write_dot(build_context, basic_conf, dot_io)
-        dot_lines = dot_io.getvalue().strip('\n').split('\n')
-        assert set(filter(lambda expected:
-                          any(filter(lambda line: re.match(expected, line),
-                                     dot_lines[1:7])),
-                          expected_dot_nodes)) == expected_dot_nodes
